Update Harbor to version 2.9.5 (#24)
This PR moves the Harbor version from version 2.7.0 to 2.9.5. In accordance to the update procedure described on the upstream website, the update covers only two minor version (2.7.0-2.9.0) and we decided to do one update at a time.

During this upgrade, both Notary and Chart Museum were removed, so please read the upgrade procedure thoroughly if you suspect this change impacts you.